pyproject.toml
[tool.poetry]
name = "pdfedit"
version = "0.1.0"
description = "Edit PDFs with google docs"
authors = ["Scoder12 <34356756+Scoder12@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.8"
google-api-python-client = "^1.10.0"
google-auth-httplib2 = "^0.0.4"
google-auth-oauthlib = "^0.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0a5"]
build-backend = "poetry.core.masonry.api"
